generate module with version info
