+++ src/cpprog/cpprog_version.cpp.in
@@ -1,0 +1,32 @@
+export module cpprog:version;
+
+import std;
+
+namespace cpprog::version {
+
+export [[nodiscard]] constexpr std::string_view as_string()
+{
+    return "${cpprog_VERSION}";
+}
+
+export [[nodiscard]] constexpr int major()
+{
+    return ${cpprog_VERSION_MAJOR};
+}
+
+export [[nodiscard]] constexpr int minor()
+{
+    return ${cpprog_VERSION_MINOR};
+}
+
+export [[nodiscard]] constexpr int patch()
+{
+    return ${cpprog_VERSION_PATCH};
+}
+
+export [[nodiscard]] constexpr std::string_view git_commit_hash()
+{
+    return "${cpprog_GIT_COMMIT_HASH}";
+}
+
+} // namespace cpprog::version
